+++ src/RaynServer/Rayn/Rayn/Controllers/HistoryController.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
-using Rayn.Models;
 using Rayn.Services.Database.Interfaces;
 using Rayn.Services.Models;
+using Rayn.ViewModels;
 
 namespace Rayn.Controllers
 {
