+++ src/RaynServer/Rayn/Rayn/ViewModels/ThreadRoomViewModel.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Rayn.Services.Realtime.Hubs;
 
-namespace Rayn.Models
+namespace Rayn.ViewModels
 {
     public class ThreadRoomViewModel
     {
