+++ src/RaynServer/Rayn/Rayn/ViewModels/ThreadViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Rayn.Models
+namespace Rayn.ViewModels
 {
     public class ThreadViewModel
     {
