+++ src/RaynServer/Rayn/Rayn/Views/ThreadRoom/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model ThreadRoomViewModel
+﻿@model Rayn.ViewModels.ThreadRoomViewModel
 @{
     Layout = "_ThreadRoomLayout";
     ViewData["ThreadTitle"] = Model.ThreadTitle;
