+++ src/RaynServer/Rayn/Rayn/Views/_ViewImports.cshtml
@@ -1,3 +1,3 @@
 ﻿@using Rayn
-@using Rayn.Models
+@using Rayn.ViewModels;
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
